requirements.txt
gunicorn==21.2.0
flask
mysql-connector-python
